fix: change min tests
